# Slightly optimise the loads in the 8x4x2 indirect BGEMM kernel