@AdamHillier 2020-10-02
> Nice. How do you feel about adding a comment to the file with something along the lines of
> 
> ```c++
> // The separated ld1 instructions are faster than the ld1r instruction,
> // see https://github.com/larq/compute-engine/pull/521
> ```

Ah yeah, will do :+1: